requirements.txt
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
click-default-group==1.2.4
exceptiongroup==1.3.1
idna==3.11
iniconfig==2.1.0
packaging==26.0
pluggy==1.6.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
-e git+https://github.com/russellbarnett/spacshortmodel.git@766af5907aebd889c1ea705ecb6f7dbed4789661#egg=slidemodel
sqlite-fts4==1.0.3
sqlite-utils==3.37
tabulate==0.9.0
tomli==2.4.0
typing_extensions==4.15.0
urllib3==1.26.20
